backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
langchain==0.1.13
langchain-openai==0.1.1
langgraph==0.0.30
faiss-cpu==1.8.0
loguru==0.7.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyPDF2==3.0.1
pandas==2.2.1
